Implemented imgui demo and bug fixes
There were bugs like the d3dcompiler not being found for some reason, plus some of the datatype mismatches were fixed so that it would stop spamming the console

+++ GDENG03/Camera.cpp
@@ -13,7 +13,7 @@ Camera::~Camera()
 	InputSystem::getInstance()->removeListener(this);
 }
 
-void Camera::update(float deltaTime)
+void Camera::update(double deltaTime)
 {
 	Vector3D localPos = this->getLocalPosition();
 	float x = localPos.getX();
@@ -22,22 +22,22 @@ void Camera::update(float deltaTime)
 	float moveSpeed = 10.0f;
 
 	if (InputSystem::getInstance()->isKeyDown('W')) {
-		z += deltaTime * moveSpeed;
+		z += (float)deltaTime * moveSpeed;
 		this->setPosition(x, y, z);
 		this->updateViewMatrix();
 	}
 	else if (InputSystem::getInstance()->isKeyDown('S')) {
-		z -= deltaTime * moveSpeed;
+		z -= (float)deltaTime * moveSpeed;
 		this->setPosition(x, y, z);
 		this->updateViewMatrix();
 	}
 	else if (InputSystem::getInstance()->isKeyDown('A')) {
-		x += deltaTime * moveSpeed;
+		x += (float)deltaTime * moveSpeed;
 		this->setPosition(x, y, z);
 		this->updateViewMatrix();
 	}
 	else if (InputSystem::getInstance()->isKeyDown('D')) {
-		x -= deltaTime * moveSpeed;
+		x -= (float)deltaTime * moveSpeed;
 		this->setPosition(x, y, z);
 		//std::cout << "x: " << x << std::endl;
 		this->updateViewMatrix();
